cleanup: zero clippy + zero rustdoc warnings; fix lib-name collision; venv simplify
- clippy --workspace --all-targets: 0 warnings (was 241). ~90 auto-fixed, ~48
  manual genuine fixes (from_ref, needless_range_loop, manual_strip, drop_non_drop
  no-ops, match_overlapping_arm made exclusive in vcs/object table.rs, MinDist
  Add impl, type aliases for handler tables), 15 targeted allows each with a
  one-line reason (TCB got attributes + comments ONLY; e.g. builtins.rs keeps
  its explicit y=0 branch).
- cargo doc --no-deps: 0 warnings (was ~269): private-link demotions, ~45 links
  retargeted to real public paths, mod@/fn@ disambiguation, redundant targets.
- crate name collision FIXED properly: covalence-python cdylib renamed to
  covalence_py; the Python module stays `import covalence` via
  #[pymodule(name = "covalence")] + maturin module-name. cargo doc now builds
  the whole workspace (no doc = false hack).
- single-checkout simplification: activate.sh transitional venv fallback removed.

Gate: bun test 15/15 (incl. full cargo suite, 2725 tests), python 531,
clippy 0, doc 0, fmt clean, serve --features static + test-guest wasm32 green.

+++ crates/kernel/hol/core/src/builtins.rs
@@ -489,6 +489,9 @@ fn int_binop(tag: IntTag, op: IntOp, ab: u64, bb: u64) -> u64 {
         // y=0, since `int.div n 0 = 0`), and `Rem`'s is `fromInt(intMod …)`
         // / `fromNat(natMod …)` (→ the dividend at y=0). `au` is the
         // dividend's low-`w` bits, i.e. its own value.
+        // TCB: keep the explicit `y = 0` branch mirroring the signed/`Rem`
+        // arms rather than restructuring into `checked_div`.
+        #[allow(clippy::manual_checked_ops)]
         IntOp::Div => {
             if tag.is_signed() {
                 let bv = value_s(tag, bb);
@@ -762,7 +765,7 @@ mod tests {
         Term::nat_lit(Nat::from_inner(n.into()))
     }
     fn int(n: i32) -> Term {
-        let nat = Nat::from_inner((n.unsigned_abs() as u32).into());
+        let nat = Nat::from_inner(n.unsigned_abs().into());
         let sign = if n == 0 {
             Sign::Zero
         } else if n > 0 {
